Admin/requirements.txt
streamlit
pypdf
langchain
faiss-cpu
boto3
langchain_community
python-dotenv